rename selector file
